Tracing layer: allow use of cached node list entries for sub-views
If a cached entry doesn't match exactly, but has a bounding box that
encloses the one of the current request and if it didn't drop any nodes
(by reaching the maximum nodes limit), the cached entry will now be used
as well. To do this, the LRUCache now allows to iterate all its items
without marking them as accessed. This makes zooming in (and back out)
faster, but causes more elements to being rendered (because *all* nodes
from a larger bounding box might be processed). To mitigate this
somewhat, the smallest cache entry that can be found is used. In
practice, this shouldn't be a problem in most situations. If it turns
out it is, this can be turned off through the Tracing Overlay section in
the Settings Widget, using the "Use cache for sub-views" setting (which
can also be set for whole instances). In my tests however, it didn't
matter for smaller and medium sized setups, but made a big difference
for large amounts of nodes and slow connections. Reloading the tracing
layer (clicking the icon) or waiting one minute will invalidate cached
entries.

A future version might be able to throw away edges that don't intersect
with the current view bounding box and make the displayed node set
smaller.