Tracing layer: make static node selection simpler and more robust
Selecting nodes explicitly with multiple stackviewers open (e.g. by
Shift-Click in the 3D Viewer) sometimes raised an error about not being
able to find the selected node. This is fixed now.

The problem before was that each tracing overlay was moved to
explicitly to the requested location and then the node was selected.
Moving to a location in an individual stack however had the side effect
of causing the whole project to relocate. Since this was done for each
layer, the project would be updated multiple times "internally" while
the outer flow control expected to have full control of the state.

Additionally, this makes node selection faster, especially with many
open stack viewers and many nodes, because less redrawing happens.