Tracing overlay: refactor connector type handling
This removes all relation specific data structures from the connector
representation on the tracing layer. Based on this, adding new connector
types and relations is much easier now. This implementation also needs
less memory and can therefore help a little bit with large fields of
view. For now, there are no performance problems, but if needed, the
performance can be improved to the expense of memory (storing indices).