Tracing layer: fix sub-view cache use with active nodes
By default, the nodes of the last ten sections are cached on the client
and used if new fields of view match the request parameters. This was
recently changed so that the cache can be used not only on exact query
parameter matches, but also if the bounding box is covered by matched
versions. This, however, didn't check if the active node changed or is
included in the cached response. This caused errors when an active node
was expected in a new section where it isn't visisble. This is fixed
now, a requirement to use a cached version is now that all extra nodes
have to match as well.