Web client: don't block the UI on network access errors
There are valid use cases to interact with the web client without
network access, e.g. when having skeletons loaded into the 3D viewer and
disconnecting from VPN. So instead of blocking the UI and don't allow
this kind of interaction, only the status bar warning and the special
handling of network access errors is kept.

Also show also the warning growl message when the network access error
is handled (not only for unhandled instances of this error.

See discussion on commit 7ae9f4d.