Add new generic connector relation: attached_to
This allows to only have a single connector linked to a treenode with
the "attached_to" relation. While this relation is very generic, it
provides flexibility in what to annotate with connector nodes. The idea
is to label the connector node to provide semantic information, e.g.
"vesicle". Using only a single connector, as opposed to linking
additional nodes to said connector, comes with the benefit of being able
to use many existing tools (e.g. the Connector Table) to view and filter
these annotations.

To create a new connector using this relation, use Alt + Click and
choose "Attachment". Attachment relation edges are displayed in an
orange color.