Landmark widget: add basic widget with semantic data support
This new widget can be opened using Ctrl + Space + "landmarks". It
allows to view and manage landmarks and landmark groups as well as their
relations. Landmarks are only semantic concepts, but can be linked to
one or more locations. This will be possible from the second tab in the
UI, but is not implemented yet.