Tracing overlay: make connector destructor more robust
When obliterate() is called on a connector node, its edges are removed
as part of the process. The way this was done before didn't test if
edges are available in the first place. To allow calling obliterate()
more than once, this check is added now, which is consistent with the
treenode implementation. Calling obliterate more than once can for
instance happen if a node is created outside of the reserved node pool
and is deleted. The deletion will cause an obliterate() call and a
section change will as well.