Connector API: make connectors/ return connectors instead of links
So far the connectors/ API returned a set of connector links connected
to a list of input skeletons. To make the API more consistent, this
endpoint is moved to connectors/links/ and the existing API is replaced
with a view that returns connector nodes directly. Given that the link
returning API is relatively new,  not used in many places and I couldn't
see any use of in RCATMAID and assume this shouldn't cause much trouble.

The new connectors/ API allows to filter connectors by their link to
particular skeletons as well as their tags. To obtain more information
on connectors, the API can also return all tags of each connector as
well as all links.

This API is needed to make connectors that use the attached_to relation
more easy to discover in separate widgets.