Landmarks: add support for linking landmarks to points
This can be done through the context menu shown when clicked on existing
links in the table or the alternative "(none)" text. This menu also
allows to delete existing links. Currently, a new location can be
created from the active node location or the current project location.

Points are not yet linked to landmark groups.