Landmarks: add support for links between landmark groups and locations
Right clicking on locations linked to landmarks opens a context menu.
This context menu now allows to add the location to landmark groups that
are conceptually linked to this landmark, if the landmark is not yet part
of it. Otherwise, it a "Remove landmark from" entry is shown.

Locations linked to a landmark group and a landmark will shown as
indices after each landmark name, just like they are shown in the
landmark table. The indices also refer to the same entry as in the
landmark table. Therefore, clicking on the same index link in either of
the landmark table or the landmarks in the group table, will move the
project location to the same location.