Tracing overlay: fix error on quick repeated deletes
Deleting many nodes quickly lead to an error. Due to new delete events
getting the active node ID on key press and not when actually called.
This is fixed now. Deletes are by default not queued if the active node
didn't change to a new node, instead a growl message is displayed.

Fixes catmaid/CATMAID#1354