3D viwer: add OBJ export support
All visible skeletons can now be exported as Wavefront OBJ file. Each
skeleton will be rendered as a separate group, colors are preserved and
written out as material file (MTL). Both files can be compressed as ZIP
file on the fly and it is possible to generate a simple line based
export as well as a mesh based representation using a user definable
radius. The generated ZIP file can be used with services like
augment.com. The resulting files can become quite large and it might be
useful to look into a CSG representation of the skeletons meshes to
get one connected mesh which then can be simplified further. To mitigate
this, the radial segment number can be lowered during export and
skeletons should be resampled.