Add memory aware LRU cache and use it for node list cache
With many nodes in one section (>50,000), the node list cache can
contribute heavily to out-of-memory errors. If there isn't much memory available in
the first place and every cached section takes more than 10MB space, it
seems better to not add things to the cache if a certain fill ratio is
reached. This is what the tracing overlay does now and uses an overall
memory fill ratio of 0.75, meaning if the available memory is used by
more than 75% no new cache entries will be added to the cache.

This works only in Chromium/Chrome, Firefox and other browsers use the
regular LRU cache.