Neuron exporter: allow meta-annotations to be used as filter
Annotations provided by the regular --required-annotation switch can
now also be meta-annotations. This is only a quick fix for the neuron
export, meta annotations are not yet exported.

Also, the internal get_annotated_entities() API needs an update to be
used in a practical way other than as implementation to a view.