Importer: add user mapping feature
The new option --map-users allows to map users referenced in the
imported data to already existing users in the database. This makes
updating existing databases easier and removes the de facto import data
requirement to contain user models.

The new --create-unknown-users option will instruct the importer to
create new inactive user accounts for users that aren't included as
objects in the imported data. If user mapping is enabled, accounts would
only be created for users that can't be mapped and are also not included
in the import data. By default --create-unknown-users is not set.