Import management command: allow import to existing data using fresh IDs
Unless --preserve-ids is set, the importer won't use the source provided
IDs anymore for storing options in the database. Instead, new objects
with new IDs are created now by default. This allows appending data from
different sources to a CATMAID instance.

To do this, all data is first reaed in, parsed and then put into an
index which links the source ID value to the deserialized object. The
objects ID field is set to None and after every object has been looked
at, all objects are saved, causing Django to create new objects. Then
all saved objects are iterated and for each object's foreign keys it is
checked if an object with this (source) ID has been saved before. If
this is the case, the new ID of this object is assigned to the foreign
key field. Otherwise, no action needs to be taken.