Project/stack import: fix host available check
Using a custom form field doesn't work, because we need to check those
fields conditionally, depending on what the import source is. Therefore,
everything is moved back into the form's clean() method. Now errors are
associated correctly with the respective field and "Unauthorized" (401)
is also viewed as a regular error. For the check, the passed in HTTP
user/pass are used and thereby tested as well.

See catmaid/CATMAID#1645