Import management command: allow duplicate class instances for neurons
When a class instance with the same name as an imported class instance
already exists, it is normally re-used. Meaning, the import data class
instance is not used, but all references to it are replace by references
to the existing class instances. The same is true for classes and
relations. Class instances of class "neuron", however, are a special
case. They represent neurons and the front-end and back-end allow two
different neurons to have the same name. Therefore, the importer now
allows this too.