Add custom swagger schema generator
This is needed to be able to read our YAML API documentation with
django-rest-swagger 2. This is from commit 65e9d2b of this repository:

https://github.com/m-haziq/django-rest-swagger-docs#integrating-django-rest-swagger

In the long term, we can think about changing our docstrings, to not
require this, but for now this seems to work. It also is less invasive
than the monkey patching we did before (to get path grouping right).

The above link says that it might be necessary to replace all APIView
occurrences with generics.GenericAPIView references if class based views
won't show their parameter fields in the documentation. This doesn't
seem to be needed in our case though, because all parameter fields show
up nicely.