Reconstruction sampler: add interval boundary handling in sampler model
Whether interval boundary nodes should be created is a propery of a
sampler should not change between domains/intervals.

This interval error field represents how much an interval can be "off"
when trying to reuse existing nodes. If boundary node creation is
enabled, boundary nodes will only be created when the interval boundary
is farther away from the closest node than this value.