Node providers: allow multiple node providers, update settings
The NODE_PROVIDER settings variable (settings.py) is replaced with the
NODE_PROVIDERS variable. The new variable takes a list of node provider
names, which are iterated as long as no result nodes are found. Replace
the former single string value with a list with this name as single
element, e.g. if the current setting reads NODE_PROVIDER = 'postgis2d',
replace it with NODE_PROVIDERS = ['postgis2d'].

Also, documentation for this option has been added.