Node back-end: improve node provider list handling
The NODE_PROVIDERS list is now iteratead correctly. Node providers need to
implement a matches() method now, which is used to test if a node provider is
valid for a particular set of parameters. And even if valid, node providers can
return None if they decide they are not useful for a particular query.

Also, af a node query now provides a "src" option, the default iteration of
node providers can be replaced by one particular node provider.