Tracing overlay: improve caching with little available memory
So far tracing data was only cached if both the available memory in
the current heap and the memory available on top of the current heap
size where below 75%. This is now extended by also requiring that, if a
value size estimate is available, this size estimate doesn't need more
than 25% of the available extra space for allocation. This is
technically not fully correct, because at the time of storage the value
is already allocated, but this heuristic is based on the memory needs
for another copy of a new element.

The motivation for this is to better deal with sections with a lot of
tracing data (>150k nodes).