Tracing overlay nodes: replace empty vis. group arrays with nulls
Since a visibility group list was created for every node, the creation
of these arrays become noticable on memory with large sets of nodes
(>100k). Allowing null to represent the fact that a node is no
visibility group might make some calling code more complex (check for
null), but saves memory. Given that this interface isn't used in many
places, this should not really be a problem at the moment.

Saving memory however is not really the main motivation for this change.
Instead it is to remove the number of garbage collection runs.