Memory aware LRU cache: ignore memory.jsHeapSizeLimit
I was under the wrong impression before that memory.jsHeapSizeLimit
represents a currently available maximum memory to which the current
heap can be extended to. It turns out this is not the case, but
memory.jsHeapSizeLimit is in fact the globally available memory for the
browser. Therefore, it is removed from the LRU memory checks.