SVG utils: fix style embedding
Recent Chrome versions seem to export style tags of XML documents with
an empty namespace attribute if no namespace is set explicitly. This
wasn't the case in the past and having this namespace attribute (xmlns)
not match the document namespace, prevents the style from being applied.

The function that is attaching styles to SVG documents makes therefore
now sure that the style namespace matches the document namespace.

Fixes catmaid/CATMAID#1690