Reconstruction sampler: support section interpolation
Some sections need to be interpolated for reasonable analysis. This can
be due to defect image data and tracing data based on it. The location
of tracing data in these sections are averaged with respect to their
parent and children. Otherwise these locations would be over-sampled.

The Sampler Widget now uses the persistent interpolatable section
entries.