3D viewer: custom label node matches override built-in label matches
If a node has multiple labels and one of them is a built-in like TODO,
the 3D Viewer did so far give precedence to the label loaded last. This
changes now so that custom labels always have precedence over others. So
even if a node is labeled with a built-in label like TODO, if it is also
tagged with a tag that matches the current custom tag pattern, the
custom label color is used.

Also, this reduces the number of label spheres that are created. Now
only one label sphere is selected to be displayed. This can be changed
easily in the future if wanted.

Fixes catmaid/CATMAID#1693