Overlay nodes: initialize all used fields in constructor
If fields are added later in the process, a new hidden class is created
by V8, which leads to the deoptimization of some functions. The
modifications allow the createVirtualNode() function to be optimized.

The optimization has been checked by looking at the output of running
Chromium like this:

chromium --no-sandbox --disk-cache-size=1 -media-cache=1 --js-flags="--trace-deopt --trace-deopt-verbose --trace-bailout" > chrome.log