Node query cache: add timestamp column and unique constraint
To be able to tell when a cache entry was added, we need a timestamp
column. We typically don't expect millions of cache entries, so the
additional data shouldn't be a practical concern. The uniqueness
constraint enforces that there can only be one cache entry per project,
depth and orientation. Doing this, makes updates easier, since they can
do an upsert.