Review back-end: optimize node count query
Using IN results also on recent Postgres versions in using the ANY
operator on the passed in list. This is much slower compared to a
regular JOIN on a typed list (which is similar in performance to using
VALUES). In my tests this cut the query time in half, which is
noticeable on larger lists.