Landmarks: add support for weighted group influence
If transformed nodes are outside of the source and target group bounding
boxes, they will now be transformed by both the regular transform and
the inverse. How much each component contributes to the final location
is weighed according to the distance of the node to the respective
groups. This allows transformations of skeletons that reach into the
target group or any space between souce and target groups.