Doc: make API documentation also available in regular docs
This uses bootprint to render our Swagger API definition to HTML. This
HTML is then included in the Sphinx documentation as a separate page. It
looks similar to the regular Swagger UI, but could still need some
improvements. The sidebar of catmaid.org does now link to the API
Reference as well.

A sphinx-doc make command "apidoc" has been added. Using "make apidoc"
will render a new version of the current API documentation to the
expected path. This is done automatically as part of the make-release
script.