3D viewer: quick fix for scale bar init in dialog embedded viewers
The scale bar expects a particular container configuration, which isn't
provided by the 3D Confirmation Dialog. To prevent this from happening
and ortho view isn't available in the dialog viewer, a disconnected DOM
element is created if the expected element isn't there. There is likely
a better solution, but I needed a quick fix. We should think about a
better solution in the future.