Neuron dendrogram: use common synapse colors for conn. partner nodes
Connector partner nodes can be displayed in the dendrogram, but so far
they all had the same color. This is updated so that each highlighted
connector partner node has a color associated matching the type of
connection it has to the connector. Presynaptic sites are shown in red,
postsynaptic sites in cyan and gap junctions in purple.