Tracing layer: try to fetch selected node if currently not available
Node limits and caching strategies can prevent nodes from showing in a
tracing layer. Selecting such a node resulted so far in an error. This
commit changes this behavior so that an attempt is made to get node
information from the back-end if it is currently not available on the
front-end.

While caching strategies should not result in an inconsistent data
display and a better querying mode needs to be found, this pragmatic fix
solves issues like simple node selection in case of missing data.

Fixes catmaid/CATMAID#1703