Annotation API: allow use of annotation names in query-targets endpoint
It is now possible to pass either IDs or annotation names for the
parameters annotated_with, not_annotated_with or sub_annotated_with. IDs
are still expected by default, but if the new "annotation_reference"
parameter is set to "name", all annotation references will be
interpreted as actual annotation names.

The main purpose of this change is to make scripting easier.