Tracing overlay: add "allow lazy node update" setting
The Tracing Overlay has a new setting "Allow lazy node updates",
available in the Settings Widget. If enabled (default), stack viewers
can skip node updates if a change didn't occur in the Viewer's field of
view. This is useful when e.g. the Connector Viewer is open while
tracing, because not all viewers get update due to the addition of a
node.