Connector API: fix error on connector type listing in non-tracing project
To be considered a tracing project, a project needs to have all required
classes and relations set up for it (which in turn is done automatically
when opening the Tracing Tool). The connector types listing returns also
relation IDs and so far raised an error when called on a non-tracing
project without the required relations. This API is called by the
Settings Widget, which displayed this error on when opened in
non-tracing projects.

Fixes catmaid/CATMAID#1708