Node backend: fix returned relation map for "used" relation map mode
This fixes a bug introduced in 23ac70f363a, found by @aschampion. The
returned relation map should return always a mapping from ID to relation
names and when all relations where returned this didn't hold anymore.
This causes all front-end arrows to be rendered as abutting connectors,
which is the default fallback.