Selection table: add option to import skeletons from CSV files
The new "Open CSV" button allows users to pick a CSV file and specify in
a new dialog how many lines to skip and which column to use as source
for skeleton IDs.

If a potential skeleton ID value contains quote characters (" and '),
they will be removed before attempting to parse the text as number. This
is useful, because some CSV exports seem to quote even numbers.