Tracing overlay: improve node deletion state update
3856f65c introduced a predictive state update when deleting nodes to not
always require a node update. Improve this so that edges are drawn
correctly and enough state is predictively updated to allow parent/child
navigation.

Note that this is still brittle when, e.g., deleting a node along an
edge then attempting to create a new child for the former parent,
because of mismatched edition times.