Skeleton API: improve performance of node count endpoint
The /{project_id}/skeleton/{skeleton_id}/node_count endpoint retrieves
the node count for a given skeleton. This API is now significantly
faster (>10x) by using the skeleton summary table to get the node count
instead of calculating it on the fly. This makes opening the Merge
Dialog faster for large skeletons, because before opening the dialog it
is checked if the merged in skeleton has only a single node (and not
show the dialog in that case).