Tracing tool: show project and stack space center when panning
So far, when panning, the status bar showed X and Y coordinates of
locations related to the panning both in project space and stack space.
The project space location reflected the mouse cursor location and the
stack space coordinates referred to the stack viewer center.

To make this more consistent with other displayed coordinates, the
displayed variables are now showing X, Y and Z coordinates both in
project space and stack space of the center of the active stack viewer.