Status bar: show mouse coords in project and stack space
There are multiple places in CATMAID where project space coordinates are
useful, for instance when defining interpolated sections in the 3D
Viewer or creating box volumes.

Also the used font for rendering is made slightly smaller and is now
colored in a dark gray rather than black, to make the added information
(project space coordinates) less prominent on the screen.