Don't handle stack loading errors multiple times
Before a parallel promise catch() call was executed after loading a
stack failed. It took care of showing the login dialog on a permission
error, but the actually returned promise was still rejected. This caused
the handler for unhandled rejected promises to be called. In case of a
permission error, it showed a second dialog besides the already
displayed one. This is fixes now and all openProjectStack() calls now
also catch rejected results.

An option has been added to handle any errors already in
openProjectStack(), which is useful when binding this method to
parameters as event handler. For a sake of cleaner API, this is not
enabled by default.