Make detailed review colors the default
This makes CATMAID display review information in 11 colors instead of 3.
Most users seem to prefer this setting, which I think justifies this
change.

To get back the old behavior, adjust the "Use detailed review colors
status" setting in the Settings Widgets. Admins can do this for whole
projects and even the whole server.