Overlay nodes: compute planeX, planeY and planeZ statically
Once a SkeletonElements object is instantiated, it is not expected to
change the tracing overlay its bound to during the course of its
lifetime. This allows us to convert planeX, planeY and planeZ from
functions to static values for the respective tracing layer.

In V8, due to the dynamic prototype construction, the previous planeX(),
planeY() and planeZ() functions were always deoptimized, because of a
changed hidden class (the underlying prototype). Optimizing this
function will probably not have a big effect though.