Stack viewer: fix reference line visibility after location flash
The general visibility accessor and the location flashing used different
means to control visibility (display = 'none' vs. visibility =
'hidden'). This has been normalized to display = 'none', because it is
more consistent with other parts of CATMAID and more applicable, because
we don't want the lines to be rendered at all when they are not shown
(rather then only being invisible).