Landmark widget: provide option to replace existing locations in edit mode
If this "Update existing landmark locations" checkbox is checked, newly
added landmark location links can replace existing location links. If
there is already a location linked to both a landmark group and a
landmark, it will be removed before the new location is added. This
allows to quickly update existing data.