Reconstruction sampler: fix silent cascaded interval deletion
The Django model configuration of sampler intervals enabled cascading
deletion of intervals when referenced treenodes are deleted. This
allowed users to delete the start node and end node or intervals while
triggering a silent deletion of the intervals. This is fixed now and the
API doesn't allow the deletion of nodes anymore that are referenced in a
sampler. On the database level this wasn't enforced and was done
entirely by Django, therefore the migration contains no actual database
changes.