Tracing layer: add optional fast merge mode
The fast merge mode setting allows users to define classes of neurons
that are merged without confirmation, just like it is done for single
node skeletons. Similar to visibility groups, this setting allows to
either have a universal match, a meta-annotation match or a match based
on the node creator.

The motivation for this is to better handle many small fragments that
are generated from a segmentation. In our case the generated skeletons
are very short and conservative about branching. Being able to quickly
merge these fragments might speed up the work with them.