Tracing overlay: fix virtual node error when loading extra nodes
If a virtual node was set as extra node to be loaded besides the regular
view, it could happen that not both his parent and his child where
available in the front-end. In this case the previous asked the back-end
for the information on the virtual node. Instead it should make sure the
child and parent of the virtual node are available and this is what is
happening now.