Layouts: add user configurable layout menu
This implements the most simple layout handling aspects: saving the
current layout, making layouts accessible in a menu, making layouts
openable. Layouts can be manually updated in the Settings Widget. Window
size ratios and window configuration is stored. A layouts is an entry of
the form

  layout("<alias>", <layout-spec>, <ratio>).

The ratio is optional.

The next step to making this useful is to also store subscription state
information.