3D viewer: prevent animation exports with odd width or height
We use the H264 coded as example in our documentation for converting the
WebM movie exported by CATMAID. H264 (and possibly other codecs) expects
the input to have a width and height of an even number. It will
otherwise. Therefore, the animation export now only allows even numbers
for width and height.