Tracing overlay: fix validity test for new connector links
This makes accidentally clicking with a connector selected less painful
when it already had a presynaptic node linked. Before, an error dialog
was shown and a node was created at the clicked location even if the
link couldn't be created. The front-end test for the validity of a new
link has now been fixed. This allows to only show a warning in this
error case and to not create the new node in the first place.