3D viewer: fix volume intersection testing
To construct volumes in THREE.js, CATMAID currently loads them as VRML
data from the server. In the recent THREE.js upgrade, we also went over
r88, which made the VRMLLoader create BufferGeometry instances. The
consequence of this is that the intersection test needs to access
individual face vertices differently.