Filters: make volume filtering work with buffer geometries
Like described in the reverted commit ea116fa3c7c, THREE.js returns now
a BufferGeometry from the VRMLLoader we use when creating volume
filters. This is now fixed by converting a passed in buffer geometry to
a regular geometry.

It would be better to expect buffer geometries and convert regular
geometries. This will be done in a separate commit at a later point int
time.