Tracing overlay: update child edit time of materialized virtual node
When a virtual node is materialized its real child node will get the new
materialized node as parent node. This update of the child node causes a
change of its edit timestamp. This edit timestamp wasn't updated in the
client after a virtual node was materialized. This had the effect hat it
wasn't possible anymore to delete the just materialized node again,
because it the wrong state would be generated for the child node (having
the wrong edit time stamp), resulting in a state matching error. This is
fixed now by explicitly updating the child node's edit timestamp,
allowing correct state generation.