GroupGraph: completed functionality for layout as columns by selections.
The "As columns" lays out the nodes of a selection as a column, sorted
by natural alphanumeric order. So e.g. a node named "LH-5" will be listed
before a node named "LH-12".
The "Fade edges" will set the opacity (as in the text field) to edges
not related to the flow of edges that passes through selected nodes.
The "Restore edges" will restore the opacity of all edges.