3D viewer: warn user before loading >100 MB image data
By default the 3D Viewer will now ask users for confirmation before
loading tiles for the Z plane display that exceed an estimated 100 MB in
total. This behavior can be deactivated by unchecking the "Data size
check" toggle in the Stack tab. If the user doesn't confirm loading that
much data, the zoom level is reset to the maximum value.

Fixes catmaid/CATMAID#1721