Tracing overlay: add client copy after node creation in event handler
When a new node is created, CATMAID avoids a full node update by
updating the stack viewer node store manually. This has been done so far
as a continuation of the node creation command. Instead of doing it
there, the "new node created" event handler is now used for this. This
increases robustness and gets rid of the second "new node created" event
that was fired previously, which in the worst case could have caused two
updates in ortho views.

This now also means that instead of potentially updating the full field
of view, when a new node is created, now either nothing is done or the
node is added to the local store.