Tracing overlay: add client copy after connector creation in event handler
Similar to treendoes, when a new connector node is created, CATMAID
avoids a full node update by updating the stack viewer node store
manually. This has been done so far as a continuation of the connector
creation command. Instead of doing it there, the "new connector created"
event handler is now used for this. This increases robustness and gets
rid of the second "new connector created" event that was fired
previously, which in the worst case could have caused two updates in
ortho views.

This now also means that instead of potentially updating the full field
of view, when a new connector is created, now either nothing is done or
the connector node is added to the local store and display.