Add option to define a skeleton cable length warning
A cable length threshold can now be defined from the Settings Widget.
If enabled, the cable length of a skeleton is checked after every
morphology change (like a join, node creation, node movement, …). If the
skeleton cable length threshold is hit, a warning is displayed. Thanks
to the skeleton summary table, the cable length lookup is fast.

See catmaid/CATMAID#1731