Options dialog: add method to add a sortable table
This method allows to pass in an array of objects, of which each needs
to have at least a "title" field.  This is array is used to populate a
table which allows row-by-row sorting or global alphanumerical sorting.
If the order is changed, the original input array is updated
accordingly.

A method like this is useful to quickly get user input on a particular
order of items without the need of additional UI in a widget.