Init: make app loading continue without superuser
CATMAID needs a superuser during normal operation. It is used to create
for instance missing classes and relations. Therefore during start-up
CATMAID checks if a superuser is available and raises an exception if
not. This causes the initialization to abort. In current Django
versions, this prevents however also the execution of the
"createsuperuser" management command. To fix this, the fact of a missing
superuser is now only printed to the log as a warning. Initialization
isn't stopped anymore.