Importer: fix test whether a project field
The test wasn't technically wrong before, but caused the ORM to try and
get the referenced project from the database. If the project doesn't
exist, this leads to an error. The purpose of this function is to
override the project to avoid this and so the test shouldn't trigger
this exception. Instead, only the ID field is checked now.