DOM: make file button reset itself when clicked
A file button is reset by settings its value to an empty string. This is
needed to open the same file again, right after it has been opened.
Without this, it isn't possible to e.g. load a JSON file in the Graph
Widget, Clear the Graph Widget and the load the same file again. With
this change, this can now be done.