Graph widget: make loading colors from JSON files more robust
In recent versions of CATMAID, colors in saved JSON files are
represented as simple numbers. These numbers are in fact a hex
representation of the color, but JSON loading code so far didn't attempt
to create colors using this hex representation. Instead, it expected an
object with the fields r, g and b. This form was common with past
THREE.js and browser versions. The color loading code in the Graph
Widget can now handle both cases to allow loading older and new JSON
files.