Export management command: only export selected fields from user model
This prevents the export of information on groups, superuser/staff
status and other potentially internal information. Not having this
information makes the import easier as well.

Unfortunately, there doesn't seem to be a better way than defining
abstract proxy models for the two options of subsets for user fields
(minimal and regular export). The 'fields' option to the serializer is
applied to all passed in models, which doesn't work with our mixed model
list.